+++ packages/terra-form-input/src/InputField.jsx
@@ -79,6 +79,10 @@ const propTypes = {
    */
   showOptional: PropTypes.bool,
   /**
+   * Specifies the type of input element to display.
+   */
+  type: PropTypes.string,
+  /**
    * The value of the input field. Use this to create a controlled input.
    */
   value: PropTypes.oneOfType([
@@ -103,6 +107,7 @@ const defaultProps = {
   refCallback: undefined,
   required: false,
   showOptional: false,
+  type: undefined,
   value: undefined,
 };
 
@@ -125,6 +130,7 @@ const InputField = (props) => {
     refCallback,
     required,
     showOptional,
+    type,
     value,
     ...customProps
   } = props;
@@ -143,6 +149,7 @@ const InputField = (props) => {
     }
   }
 
+  const inputType = type || inputAttrs.type;
   return (
     <Field
       label={label}
@@ -163,6 +170,7 @@ const InputField = (props) => {
       <Input
         {...inputAttrs}
         id={inputId}
+        type={inputType}
         onChange={onChange}
         value={value}
         defaultValue={defaultValue}