+++ packages/terra-form-input/src/terra-dev-site/doc/common/NumberInputExample.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Field from 'terra-form-field';
+import Input from '../../../Input';
+
+const NumberInputExample = () => (
+  <div>
+    <Field label="Numeric Input" htmlFor="numeric">
+      <Input name="number input" placeholder="enter digits" id="numeric" type="number" />
+    </Field>
+  </div>
+);
+
+export default NumberInputExample;